# internal/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Arduino-App-Cli
  version: 0.1.0
servers:
- description: local server
  url: http://localhost:8800
...
paths:
...
  /v1/apps/{id}:
    delete:
      description: Remove the given app and all the resources it created
      operationId: deleteApp
      parameters:
      - description: application identifier.
        in: path
        name: id
        required: true
        schema:
          description: application identifier.
          type: string
      responses:
        "200":
          description: Successful response
        "400":
          $ref: '#/components/responses/BadRequest'
        "412":
          $ref: '#/components/responses/PreconditionFailed'
        "500":
          $ref: '#/components/responses/InternalServerError'
      summary: delete the app
      tags:
      - Application
    get:
      description: Return all the detail for the given app
      operationId: getAppDetails
      parameters:
      - description: application identifier.
        in: path
        name: id
        required: true
        schema:
          description: application identifier.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppDetailedInfo'
          description: Successful response
        "412":
          $ref: '#/components/responses/PreconditionFailed'
        "500":
          $ref: '#/components/responses/InternalServerError'
      summary: Get app/example detail
      tags:
      - Application
    patch:
      description: Edit the given application. Is it possible to modify the default
        status, to add/remove/update bricks and bricks variables.
      operationId: editApp
      parameters:
      - description: application identifier.
        in: path
        name: id
        required: true
        schema:
          description: application identifier.
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppDetailedInfo'
          description: Successful response
        "400":
          $ref: '#/components/responses/BadRequest'
        "412":
          $ref: '#/components/responses/PreconditionFailed'
        "500":
          $ref: '#/components/responses/InternalServerError'
      summary: Update App Details
      tags:
      - Application
...
  /v1/bricks:
    get:
      description: Returns all the existing bricks. Bricks that are ready to use are
        marked as installed.
      operationId: getBricks
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrickListResult'
          description: Successful response
        "500":
          $ref: '#/components/responses/InternalServerError'
      summary: Get a list of available bricks
      tags:
      - Brick
...
  /v1/models/ei/projects/{projectID}:
    put:
      description: Download and install a custom Edge Impulse AI model using the provided
        project API key.
      operationId: installEIModel
      parameters:
      - description: Edge Impulse project ID
        in: path
        name: projectID
        required: true
        schema:
          description: Edge Impulse project ID
          example: 123456
          type: integer
      - description: Edge Impulse project API key
        in: header
        name: x-api-key
        required: true
        schema:
          description: Edge Impulse project API key
          example: your_edge_impulse_api_token
          type: string
      requestBody:
        content:
          application/json:
            schema:
              properties:
                impulse_id:
                  description: Edge Impulse impulse ID
                  example: 1
                  type: integer
              required:
              - impulse_id
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIModelItem'
          description: Successful response
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "500":
          $ref: '#/components/responses/InternalServerError'
        "507":
          $ref: '#/components/responses/InsufficientStorage'
      summary: Download and install a custom Edge Impulse AI model
      tags:
      - AIModels
...
components:
  responses:
    BadRequest:
      content:
        application/json:
          example:
            details: The request is invalid or missing required parameters.
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      description: Bad Request
...
    Forbidden:
      content:
        application/json:
          example:
            details: You do not have permission to access this resource.
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      description: Forbidden
    InsufficientStorage:
      content:
        application/json:
          example:
            details: Insufficient storage to complete the request.
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      description: Insufficient Storage
    InternalServerError:
      content:
        application/json:
          example:
            details: An unexpected error occurred.
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      description: Internal Server Error
...
    PreconditionFailed:
      content:
        application/json:
          example:
            details: The request is invalid.
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      description: Precondition Failed
    Unauthorized:
      content:
        application/json:
          example:
            details: Unauthorized access to the resource.
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      description: Unauthorized
  schemas:
...
    AIModelItem:
      properties:
        brick_ids:
          items:
            type: string
          nullable: true
          type: array
        description:
          type: string
        id:
          example: bGxhbWFjcHA6Z2VtbWEtMy0xYi1pdC1RNF8w
          type: string
        id_decoded:
          example: llamacpp:gemma-3-1b-it-Q4_0
          type: string
        is_builtin:
          type: boolean
        metadata:
          additionalProperties:
            type: string
          type: object
        name:
          type: string
        origin:
          $ref: '#/components/schemas/ModelOrigin'
        runner:
          type: string
        size:
          minimum: 0
          nullable: true
          type: integer
        status:
          $ref: '#/components/schemas/ModelStatus'
      type: object
...
    AppDetailedBrick:
      properties:
        category:
          type: string
        id:
          type: string
        name:
          type: string
        require_model:
          type: boolean
      required:
      - id
      - name
      type: object
    AppDetailedInfo:
      properties:
        bricks:
          items:
            $ref: '#/components/schemas/AppDetailedBrick'
          type: array
        default:
          type: boolean
        description:
          type: string
        example:
          type: boolean
        icon:
          type: string
        id:
          type: string
        name:
          type: string
        path:
          type: string
        status:
          $ref: '#/components/schemas/Status'
      required:
      - id
      - name
      - status
      type: object
...
    BrickListItem:
      properties:
        author:
          type: string
        category:
          type: string
        description:
          type: string
        id:
          type: string
        name:
          type: string
        require_model:
          type: boolean
        status:
          type: string
      type: object
    BrickListResult:
      properties:
        bricks:
          items:
            $ref: '#/components/schemas/BrickListItem'
          nullable: true
          type: array
      type: object
...
    EditRequest:
      properties:
        default:
          nullable: true
          type: boolean
        description:
          description: application description
          example: This is my awesome app
          nullable: true
          type: string
        icon:
          description: application icon
          example: "\U0001F4BB"
          nullable: true
          type: string
        name:
          description: application name
          example: My Awesome App
          nullable: true
          type: string
      type: object
    ErrorResponse:
      properties:
        code:
          type: integer
        message:
          type: string
      type: object
...
    ModelOrigin:
      description: 'Where the model came from: "curated" is declared by the internal
        model list and installs from its id alone, "user" was downloaded from a source
        the caller supplied and needs that source again, "edge-impulse-user-project"
        was deployed from the caller''s own Edge Impulse project.'
      enum:
      - curated
      - user
      - edge-impulse-user-project
      type: string
      uniqueItems: true
    ModelStatus:
      description: Model status
      enum:
      - installed
      - not-installed
      - downloading
      type: string
      uniqueItems: true
...
    Status:
      description: Application status
      enum:
      - starting
      - running
      - stopping
      - stopped
      - failed
      - uninitialized
      type: string
      uniqueItems: true
...